Create zombieburning.ptpl
Using the Zombie AI base poses the problem of your mobs spawning as babies without model differences, creating incredibly janky hitboxes and animations. Instead, just make your entity target both players and Villagers, and plug this procedure into your mob's 'On entity tick update' trigger, and you'll have them roasting in the sunlight in no time!

The logic goes as follows: 'If it's daytime in the world, check that the entity is under the sky where the sun would be, and make sure it isn't in water either. If that's the case, set the entity on fire for 10 seconds (Though this will effectively be endless if they stay under the sun.)'